Fixed bug in Orientation Handler
Fixed bug where pitch/roll angle was changed despite only moving along the Yaw axis.
For example:
Assume the quadcopter is level. Rotate it 90 degrees in the y (roll) axis. Rotate it 90 degrees in the x (pitch) axis. The pitch angle should still be 0 degrees and the roll angle should still be 90 degrees. Before this fix, the pitch angle would be at 90 degrees.