Removed need to calibrate in ESC. Cycle time is now 16ms
Turns out the reason why the quadcopter was acting strangely was because of a bad battery. Random sudden losses and regains of power made the ESCs think they were receiving pulses.

The ECSs now receive pulses every 16ms. All of the control code is still updating on a 4ms cycle though. The ESCs can receive pulses up to about 20ms apart instead of just 4ms.

ESC calibration is now done with the signals coming from the remote. Before starting the quadcopter the remote must be on and in the full throttle position. After a few beeps the throttle is put to the zero position. After another beep the quadcopter should be ready to fly.